Fix problem with dir name
Entry name can be in different forms:
  [1] index.org
  [2] content/index.org
  [3] content/daily/index.org
  [4] content/dir1/dir2/index.org

"file-name-directory" returns nil for [1], "content/" for [2], and
"content/.../" for the rest.  The prefix "content/" needs to be removed
first, so after removal it becomes either an empty string, or ".../".
Finally, remove the "/" suffix.

The complete process: get dir name, remove the "content/" prefix, and
remove the "/" suffix.  To demonstrate:
[1] index.org -> nil
[2] content/index.org -> content/ -> ""
[3] content/daily/index.org -> content/daily/ -> daily/ -> daily
[4] content/dir1/dir2/index.org -> content/dir1/dir2/ -> dir1/dir2/ -> dir1/dir2